Package the files
