+++ src/main/java/duke/Duke.java
@@ -1,4 +1,4 @@
-import java.text.spi.NumberFormatProvider;
+package duke;
 import java.util.Scanner;
 import java.util.ArrayList;
